fix packages structure
