+++ src/test/java/com/hackathon/testproject/pages/nyse/QuotePage.java
@@ -1,0 +1,7 @@
+package com.hackathon.testproject.pages.nyse;
+
+import com.hackathon.testproject.pages.AbstractPage;
+
+public class QuotePage extends AbstractPage {
+
+}
