+++ src/test/java/com/hackathon/testproject/steps/NyseSteps.java
@@ -1,0 +1,11 @@
+package com.hackathon.testproject.steps;
+
+import static com.hackathon.testproject.steps.BaseSteps.driver;
+
+public class NyseSteps  {
+
+    public static void navigateToSite() {
+        driver().navigate().to("https://www.nyse.com/");
+    }
+
+}
