--- src/test/java/pages/nyse/QuotePage.java
@@ -1,5 +1,0 @@
-package pages.nyse;
-
-public class QuotePage {
-
-}
