--- src/test/java/steps/NyseSteps.java
@@ -1,6 +1,0 @@
-package steps;
-
-public class NyseSteps  {
-
-
-}
