--- src/test/java/ui/Scenario2Test.java
@@ -1,11 +1,0 @@
-package ui;
-
-import org.testng.annotations.Test;
-
-public class Scenario2Test {
-
-    @Test
-    public void test() {
-
-    }
-}
